+++ src/commands/progress.ts
@@ -17,7 +17,7 @@ import * as barHorizontal from 'bar-horizontal';
 // --------------------------------------------------
 export default function progress( INPUT, ARGS, config ) {
 	return new Promise( ( resolve, reject ) => {
-		let log = config.utils.getLog( 'active' );
+		let log = ARGS.archive ? config.utils.getLog( 'archive' ) : config.utils.getLog( 'active' );
 		let { goals } = log;
 
 		let total = Object.keys( goals ).length;
@@ -30,10 +30,11 @@ export default function progress( INPUT, ARGS, config ) {
 		config.debugger.log( `Incomplete: ${incomplete}\n` );
 
 		// Only display bar chars when:
+		// - goal data exist;
 		// - command was executed from CLI;
 		// - Goalist is not running in 'silent' mode.
-		if ( config.cli && config.debugger.getMode() !== 'silent' ) {
-			barHorizontal(
+		if ( total && config.cli && config.debugger.getMode() !== 'silent' ) {
+			config.utils.barHorizontal(
 				{
 					'Complete': complete,
 					'Incomplete': incomplete,
@@ -45,7 +46,7 @@ export default function progress( INPUT, ARGS, config ) {
 		}
 
 		resolve( {
-			type: 'active', /// TODO: Ensure that `type` value reflects 'active' or 'archived' progress.
+			type: ARGS.archive ? 'archive' : 'active',
 			total,
 			complete,
 			incomplete,